% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxTree.R
\name{checkboxTreeInput}
\alias{checkboxTreeInput}
\title{<Add Title>}
\usage{
checkboxTreeInput(
  inputId,
  nodes,
  sort = FALSE,
  checkModel = "leaf",
  checked = list(),
  onlyLeafCheckboxes = FALSE,
  showExpandAll = FALSE
)
}
\description{
<Add Description>
}
